requirements.txt
flask==3.0.0
flask-cors==4.0.0
faiss-cpu==1.7.4
numpy>=1.24.0,<2.0.0
sentence-transformers>=2.3.0
torch>=2.1.0
huggingface_hub>=0.20.0
transformers>=4.30.0
zhipuai>=2.0.0

